Fix bug - turns out the CSS to use a cool font was never included :(
I downloaded Poppins, but I forgot to use @font-face to load it. The entirety of Clear Crystal Conundrums uses Arial.
Because Poppins is slightly wider than Arial, I'd have to redo all my line breaks if I wanted to use it, and I don't especially want to. It's good with Arial.